# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auto-Harness Optimization Service",
    "description": "Backend service that runs the TerminalBench agent against a task subset, observes failures,\nand (optimize mode) uses an LLM to iteratively improve the agent \u2014 proposing a new `agent.py`,\nrunning it in an isolated **E2B sandbox**, and keeping the change only if it beats the best\nscore seen so far.\n\n**Auth:** all `/v1` endpoints require an `X-API-Key` header (seeded dev key: `dev-key`).\n\n**Job lifecycle:** `POST /v1/jobs` returns immediately with a `queued` job; a background worker\nclaims it, runs it, and persists the full, lossless iteration history (agent source, per-task\nrewards + traces, the LLM's proposal request/response). Poll `GET /v1/jobs/{id}` until the\nstatus is terminal, then read `GET /v1/jobs/{id}/iterations` for the complete trajectory.\n",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Liveness + DB/worker readiness",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit a benchmark/optimize job (returns immediately)",
        "operationId": "submit_job_v1_jobs_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List jobs in your org",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItem"
                  },
                  "title": "Response List Jobs V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get job status + structured result summary",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/iterations": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Full lossless iteration history (incl. LLM proposal audit trail)",
        "operationId": "get_iterations_v1_jobs__job_id__iterations_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IterationRead"
                  },
                  "title": "Response Get Iterations V1 Jobs  Job Id  Iterations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExecutorKind": {
        "type": "string",
        "enum": [
          "simulated",
          "harbor"
        ],
        "title": "ExecutorKind"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "0.1.0"
          },
          "db": {
            "type": "string",
            "title": "Db"
          },
          "worker_enabled": {
            "type": "boolean",
            "title": "Worker Enabled"
          }
        },
        "type": "object",
        "required": [
          "db",
          "worker_enabled"
        ],
        "title": "HealthResponse"
      },
      "IterationDecision": {
        "type": "string",
        "enum": [
          "baseline",
          "accepted",
          "rejected",
          "error"
        ],
        "title": "IterationDecision"
      },
      "IterationRead": {
        "properties": {
          "idx": {
            "type": "integer",
            "title": "Idx"
          },
          "decision": {
            "$ref": "#/components/schemas/IterationDecision"
          },
          "decision_reason": {
            "type": "string",
            "title": "Decision Reason"
          },
          "val_score": {
            "type": "number",
            "title": "Val Score"
          },
          "n_passed": {
            "type": "integer",
            "title": "N Passed"
          },
          "n_failed": {
            "type": "integer",
            "title": "N Failed"
          },
          "agent_hash": {
            "type": "string",
            "title": "Agent Hash"
          },
          "agent_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Agent Params"
          },
          "proposer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposerKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "proposal_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Rationale"
          },
          "proposal_diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Diff"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "task_results": {
            "items": {
              "$ref": "#/components/schemas/TaskResultRead"
            },
            "type": "array",
            "title": "Task Results",
            "default": []
          },
          "agent_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Source"
          },
          "llm_request": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Request"
          },
          "llm_response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Response"
          }
        },
        "type": "object",
        "required": [
          "idx",
          "decision",
          "decision_reason",
          "val_score",
          "n_passed",
          "n_failed",
          "agent_hash",
          "agent_params",
          "created_at"
        ],
        "title": "IterationRead"
      },
      "JobCreate": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/JobMode",
            "default": "single_run"
          },
          "executor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutorKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "subset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Subset",
            "default": "core"
          },
          "max_iterations": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Iterations"
          },
          "patience": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Patience"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "title": "JobCreate"
      },
      "JobListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "mode": {
            "$ref": "#/components/schemas/JobMode"
          },
          "executor": {
            "$ref": "#/components/schemas/ExecutorKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "best_val_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Val Score"
          },
          "n_iterations": {
            "type": "integer",
            "title": "N Iterations"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mode",
          "executor",
          "status",
          "n_iterations",
          "created_at"
        ],
        "title": "JobListItem"
      },
      "JobMode": {
        "type": "string",
        "enum": [
          "single_run",
          "optimize"
        ],
        "title": "JobMode"
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "mode": {
            "$ref": "#/components/schemas/JobMode"
          },
          "executor": {
            "$ref": "#/components/schemas/ExecutorKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "subset": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subset"
          },
          "max_iterations": {
            "type": "integer",
            "title": "Max Iterations"
          },
          "patience": {
            "type": "integer",
            "title": "Patience"
          },
          "n_iterations": {
            "type": "integer",
            "title": "N Iterations"
          },
          "best_val_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Val Score"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "baseline_val_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Val Score"
          },
          "train_subset": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Train Subset"
          },
          "test_subset": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Subset"
          },
          "test_val_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Val Score"
          },
          "improvement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improvement"
          }
        },
        "type": "object",
        "required": [
          "id",
          "org_id",
          "user_id",
          "mode",
          "executor",
          "status",
          "subset",
          "max_iterations",
          "patience",
          "n_iterations",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
      "JobSummary": {
        "properties": {
          "val_score": {
            "type": "number",
            "title": "Val Score"
          },
          "n_passed": {
            "type": "integer",
            "title": "N Passed"
          },
          "n_failed": {
            "type": "integer",
            "title": "N Failed"
          },
          "passed_tasks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Passed Tasks"
          },
          "failed_tasks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed Tasks"
          },
          "failures": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Failures"
          }
        },
        "type": "object",
        "required": [
          "val_score",
          "n_passed",
          "n_failed",
          "passed_tasks",
          "failed_tasks",
          "failures"
        ],
        "title": "JobSummary",
        "description": "Structured result of the best iteration (M1 = the baseline iteration)."
      },
      "ProposerKind": {
        "type": "string",
        "enum": [
          "openai",
          "mock"
        ],
        "title": "ProposerKind"
      },
      "TaskResultRead": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "reward": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reward"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "trace_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Excerpt"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "reward",
          "passed"
        ],
        "title": "TaskResultRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}